Added basic tile-based movement
